Major performance improvements
Changes:
 - Added several typedefs to make code cleaner
 - Large data structures are always passed by reference or const reference (not by pointer)
 - PhotonCount is now much smarter about the range of times it tracks, which saves time during the noise addition and clearing phases
 - Using unsigned long for size type instead of int
 - The Plane and Vector classes now throw runtime errors if they are given invalid input, which prevents unexpected/undefined behavior
 - The user can now decide, via configuration file, whether they want the rng to be randomly seeded
 - Changed the noise clearing and cluster search methods to use iterative instead of recursive traversals, which is both faster and will prevent possible stack overflows
 - The positions of rays are now jittered (previously, only times were)
 - Got rid of the Reconstructor.ThreeSigmaFilter method, as it wasn't doing anything not already done by RecursiveSearch
 - Roots of quadratics are calculated manually instead of using the ROOT Polynomial, which was inefficient

+++ cherenkov_lib/Utility.cpp
@@ -21,7 +21,7 @@ namespace cherenkov_simulator
     TVector3 Utility::ToVector(string s)
     {
         // Clear out everything before the first parenthesis.
-        int current = s.find('(');
+        unsigned long current = s.find('(');
         s.erase(0, current + 1);
 
         // Pull double values from the vector.
